backend/requirements.txt
# Azure Security Platform V2 - Backend Dependencies
# Python 3.11+

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
gunicorn==21.2.0
python-multipart==0.0.9

# Azure SDKs
azure-identity==1.15.0
azure-mgmt-core==1.4.0
azure-mgmt-recoveryservices==2.5.0
azure-mgmt-recoveryservicesbackup==9.1.0
azure-cosmos==4.5.1
azure-keyvault-secrets==4.7.0

# Microsoft Graph
msgraph-sdk==1.2.0
microsoft-kiota-authentication-azure==1.0.0

# Async & HTTP
httpx==0.26.0
aiohttp==3.9.3

# Caching
redis==5.0.1

# Scheduling
apscheduler==3.10.4

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic>=2.6
pydantic-settings>=2.2

# PDF Generation
reportlab==4.1.0

# Utilities
python-dotenv==1.0.1
structlog==24.1.0

# Development & Testing
pytest>=7.0.0,<8.0.0
pytest-asyncio==0.23.4
